move AppConfig to an own file
